refactor: service 객체 제거 및 game 도메인 객체 생성
